NXT version partially working. No USB. Config stuff mainly in target files. Run out of time.

+++ UAVXArm32F4/src/misctypes.h
@@ -36,7 +36,7 @@ typedef double real64;
 typedef void (*voidFuncPtr)(void);
 typedef void (*i16FuncPtr)(int16 * v);
 typedef unsigned char BYTE;
-typedef int idx; // mainly short structure/loop indexing - was uint8 but Arm is native 32
+typedef uint_fast8_t idx; // mainly short structure/loop indexing - was uint8 but Arm is native 32
 
 /*
  typedef struct {
